refactor(utxo-sql): remove unlockBatcher single-hash SetLocked fan-out
Deletes the per-item batcher that was routing single-hash SetLocked(false)
calls through a done-channel fan-out. setUnlockedBulk already handles
arbitrary slice sizes including len=1, so the dedicated batcher-plus-done-
channel path was adding goroutines, channel ops, and scheduler overhead
for no gain.

Removes:
- Store.unlockBatcher field
- batchUnlockItem type
- unlockBatcher initialisation block
- single-hash branch in SetLocked
- sendUnlockBatch callback
- the unlock-batcher-specific test cases

Keeps the Postgres testcontainer setup helper under a new name
(postgres_testcontainer_setup_test.go) because spend_dah_test.go depends
on it. Settings LockedBatcherSize / LockedBatcherDurationMillis are
preserved because the aerospike store still uses them.
