+++ stores/utxo/sql/postgres_testcontainer_setup_test.go
@@ -1,0 +1,56 @@
+package sql
+
+import (
+	"context"
+	"net/url"
+	"testing"
+	"time"
+
+	"github.com/bsv-blockchain/teranode/ulogger"
+	"github.com/bsv-blockchain/teranode/util/test"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"github.com/testcontainers/testcontainers-go"
+	"github.com/testcontainers/testcontainers-go/modules/postgres"
+	"github.com/testcontainers/testcontainers-go/wait"
+)
+
+// setupPostgresStore starts a Postgres testcontainer and returns a *Store
+// suitable for integration tests. Preserves its historical name (used by
+// other _test.go files in this package) even though the unlock-batcher
+// test cases it was originally designed for are gone.
+func setupPostgresStore(t *testing.T) (*Store, context.Context) {
+	t.Helper()
+	ctx := context.Background()
+
+	pgContainer, err := postgres.Run(ctx,
+		"postgres:13",
+		postgres.WithDatabase("testdb"),
+		postgres.WithUsername("testuser"),
+		postgres.WithPassword("testpass"),
+		testcontainers.WithWaitStrategy(
+			wait.ForLog("database system is ready to accept connections").
+				WithOccurrence(2).
+				WithStartupTimeout(5*time.Minute),
+		),
+	)
+	require.NoError(t, err)
+	t.Cleanup(func() {
+		assert.NoError(t, pgContainer.Terminate(ctx))
+	})
+
+	connStr, err := pgContainer.ConnectionString(ctx, "sslmode=disable")
+	require.NoError(t, err)
+
+	dbURL, err := url.Parse(connStr)
+	require.NoError(t, err)
+
+	tSettings := test.CreateBaseTestSettings(t)
+	tSettings.UtxoStore.DBTimeout = 30 * time.Second
+	tSettings.BatcherDrainMode = true // batcher fires immediately in tests
+
+	store, err := New(ctx, ulogger.TestLogger{}, tSettings, dbURL)
+	require.NoError(t, err)
+
+	return store, ctx
+}
