+++ stores/utxo/sql/sql.go
@@ -103,13 +103,6 @@ type Store struct {
 	spendBatcher    *batcher.Batcher[batchSpend]
 	getBatcher      *batcher.Batcher[batchGetItem]
 	createBatcher   *batcher.Batcher[batchCreateItem]
-	unlockBatcher   *batcher.Batcher[batchUnlockItem]
-}
-
-// batchUnlockItem represents a single SetLocked(false) request.
-type batchUnlockItem struct {
-	hash chainhash.Hash
-	done chan error
 }
 
 // batchGetItemData holds the result of a batch get operation.
@@ -225,16 +218,6 @@ func New(ctx context.Context, logger ulogger.Logger, tSettings *settings.Setting
 		s.createBatcher = batcher.New(storeBatchSize, storeBatchDuration, s.sendCreateBatch, true)
 		if tSettings.BatcherDrainMode {
 			s.createBatcher.SetDrainMode(true)
-		}
-	}
-
-	// Initialize unlock batcher for Postgres — batches single-hash SetLocked(false) calls.
-	if storeURL.Scheme == "postgres" && tSettings.UtxoStore.LockedBatcherSize > 1 {
-		unlockBatchSize := tSettings.UtxoStore.LockedBatcherSize
-		unlockBatchDuration := time.Duration(tSettings.UtxoStore.LockedBatcherDurationMillis) * time.Millisecond
-		s.unlockBatcher = batcher.New(unlockBatchSize, unlockBatchDuration, s.sendUnlockBatch, true)
-		if tSettings.BatcherDrainMode {
-			s.unlockBatcher.SetDrainMode(true)
 		}
 	}
 
@@ -4110,23 +4093,8 @@ func (s *Store) SetLocked(ctx context.Context, txHashes []chainhash.Hash, setVal
 		return nil
 	}
 
-	// Postgres: single-hash unlock → use batcher.
-	if s.unlockBatcher != nil && len(txHashes) == 1 {
-		if ctx.Err() != nil {
-			return ctx.Err()
-		}
-		done := make(chan error, 1)
-		s.unlockBatcher.Put(&batchUnlockItem{hash: txHashes[0], done: done})
-		select {
-		case err := <-done:
-			return err
-		case <-ctx.Done():
-			s.logger.Warnf("[setLockedBatched] context cancelled while waiting for batcher result — unlock may or may not be applied")
-			return ctx.Err()
-		}
-	}
-
 	// Postgres: bulk unlock + DAH in chunked UPDATE statements
+	// (handles len=1 as well; no dedicated single-hash path).
 	if s.engine == "postgres" {
 		return s.setUnlockedBulk(ctx, txHashes)
 	}
@@ -5109,26 +5077,4 @@ func isLockError(err error) bool {
 	return strings.Contains(errStr, "database is locked") ||
 		strings.Contains(errStr, "deadlock") ||
 		strings.Contains(errStr, "lock timeout")
-}
-
-// sendUnlockBatch processes a batch of SetLocked(false) calls via setUnlockedBulk,
-// which chunks large batches into multiple UPDATE statements (maxINClauseSize=400).
-func (s *Store) sendUnlockBatch(batch []*batchUnlockItem) {
-	ctx := s.ctx
-
-	if len(batch) == 1 {
-		hashes := []chainhash.Hash{batch[0].hash}
-		batch[0].done <- s.setUnlockedBulk(ctx, hashes)
-		return
-	}
-
-	hashes := make([]chainhash.Hash, len(batch))
-	for i, item := range batch {
-		hashes[i] = item.hash
-	}
-
-	err := s.setUnlockedBulk(ctx, hashes)
-	for _, item := range batch {
-		item.done <- err
-	}
 }
